Use 10ms (vs 1s) as a time for setting scrypt work
Spartacus uses a password-based rage file as an encryption layer around a
simple file database format. We do this partly to minimize the time that
important secrets are held in memory. Whenever a user mutates the
database, we pull an encryption key from the system keychain (which on
most platforms involves verifying the user), and use that to read and
write from the age-encrypted file.

The key we use (and store in the keychain) is a long randomly-generated
alphanumeric string (on most platforms, at least; and on platforms with
no system keychain, it's the empty string). Thus, stretching the
password is unnecessary, and makes the user experience substantially
worse.

There has been talk of custom work factors (e.g. in a75275), but it was
discarded because of a pending refactor about a year ago. I don't care
about configurability for Spartacus; I just care about having a sane
value. So this commit patches rage with a 10ms-targeting work factor.